Split post 1423 oversized slices
