+++ app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
@@ -577,571 +577,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
     // tests are deleted with the feature. The conversation feed now shows every
     // event with no query filter. ─────────────
 
-    @Test
-    fun stoppedAgentLogTailMarksConversationStaleWhileTmuxStaysConnected() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        val tailJob = Job()
-        vm.startAgentConversationForTest("%0", detection)
-
-        val started = vm.startAgentTailForTest(
-            paneId = "%0",
-            session = FakeSshSession(tailJob = tailJob),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-
-        // Issue #576: the VM now wires the batched tail, so the returned Job is
-        // the umbrella that owns the underlying SSH tail + drain (not the raw
-        // tail job). The contract the test cares about is that a tail started
-        // and its terminal cause still drives the conversation sync status.
-        assertNotNull(started)
-        assertEquals(AgentConversationSyncStatus.Live, vm.agentConversations.value["%0"]!!.syncStatus)
-
-        tailJob.complete()
-        advanceUntilIdle()
-
-        assertEquals(
-            "normal tail exit means the conversation feed is stale, not that tmux disconnected",
-            AgentConversationSyncStatus.Stale,
-            vm.agentConversations.value["%0"]!!.syncStatus,
-        )
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-    }
-
-    @Test
-    fun codexNewReplayBurstCoalescesIntoFewStateFlowEmissionsNotPerLine() = runTest(scheduler) {
-        // Issue #576 (the load-bearing regression): a Codex `/new` rewrites
-        // the rollout JSONL with thousands of lines. With the per-line tail
-        // (the old shape) each parsed event drove its own
-        // `appendAgentEvents` → `updateAgentConversation` → ONE
-        // `_agentConversations` StateFlow emission, so an N-line replay fired
-        // N emissions → N O(n) conversation re-derivations on the main thread
-        // (the recompose storm the maintainer saw). The VM now wires
-        // `tailEventsBatchedFromLine`, so the burst must collapse to a HANDFUL
-        // of emissions, NOT one per line — while still ingesting every event.
-        val lineCount = 2_000
-        val replay = (0 until lineCount).map { index ->
-            """{"type":"user","uuid":"u$index","message":{"role":"user","content":"replayed $index"}}"""
-        }
-        // The repository's tail drain runs on `tailScope`; wiring it to this
-        // test's virtual clock makes the 50 ms debounce window deterministic.
-        val repository = AgentConversationRepository(
-            tailScope = backgroundScope,
-            tailBatchWindowMillis = 50L,
-        )
-        val vm = newVm(agentRepository = repository)
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        vm.startAgentConversationForTest("%0", detection)
-
-        // Count distinct `agentConversations` emissions for the pane: each
-        // emission is one reconcile + one recompose trigger. A per-line tail
-        // would produce ~`lineCount` of these; the batched tail a handful.
-        var emissionCount = 0
-        var lastEventCount = 0
-        val collector = backgroundScope.launch {
-            vm.agentConversations.collect { map ->
-                map["%0"]?.let {
-                    emissionCount += 1
-                    lastEventCount = it.events.size
-                }
-            }
-        }
-        runCurrent()
-        val emissionsAfterRegister = emissionCount
-
-        vm.startAgentTailForTest(
-            paneId = "%0",
-            session = ReplayTailSshSession(replayLines = replay),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-        // Let the drain coroutine fire its debounce window(s). The whole
-        // replay arrives in one synchronous burst from the fake tail, so it
-        // coalesces into a single batch → a single append → a single emission.
-        advanceTimeBy(500L)
-        runCurrent()
-        collector.cancel()
-
-        val burstEmissions = emissionCount - emissionsAfterRegister
-        // The headline assertion: O(batches), NOT O(N). A per-line tail would
-        // have produced ~2000 burst emissions; the batched tail a tiny number.
-        assertTrue(
-            "expected a handful of burst emissions, got $burstEmissions for $lineCount replayed lines",
-            burstEmissions in 1..5,
-        )
-        // ...and the conversation is still correct: every event was ingested,
-        // then bounded to the most-recent window (the same final feed a
-        // per-line ingest would yield — reconcile is order-preserving and the
-        // bound trims the oldest). The most recent replayed id must survive.
-        assertEquals(
-            "batched ingest must produce the bounded feed, not drop to empty",
-            500,
-            lastEventCount,
-        )
-        val finalEvents = vm.agentConversations.value["%0"]!!.events
-        assertEquals(500, finalEvents.size)
-        assertEquals(
-            "the most-recent replayed event must be present after coalesced ingest",
-            "u${lineCount - 1}",
-            finalEvents.last().id,
-        )
-        assertEquals(
-            "the oldest surviving event is the start of the bounded window",
-            "u${lineCount - 500}",
-            finalEvents.first().id,
-        )
-    }
-
-    @Test
-    fun stoppedAgentLogTailPreservesConcurrentConversationState() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        val tailJob = Job()
-        vm.startAgentConversationForTest("%0", detection)
-        vm.startAgentTailForTest(
-            paneId = "%0",
-            session = FakeSshSession(tailJob = tailJob),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        vm.appendAgentEventsForTest(
-            "%0",
-            listOf(
-                ConversationEvent.Message(
-                    id = "assistant-late",
-                    agent = AgentKind.ClaudeCode,
-                    atMillis = 10L,
-                    role = ConversationRole.Assistant,
-                    text = "late event",
-                ),
-            ),
-        )
-        tailJob.complete()
-        advanceUntilIdle()
-
-        val after = vm.agentConversations.value["%0"]!!
-        assertEquals(AgentConversationSyncStatus.Stale, after.syncStatus)
-        assertEquals(SessionTab.Conversation, after.selectedTab)
-        assertEquals("assistant-late", after.events.single().id)
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-    }
-
-    // Issue #1168: durable anchor for the tmux TestMainDispatcher flake
-    // (blocked PRs #1211/#1217). The AgentConversationRepository tail drain
-    // (`tailScope.launch { while (isActive) delay() }`) runs on the injected
-    // real-`Dispatchers.IO` `tailScope`. The default `newVm()` repo USED to be
-    // built with the repository's OWN default real-IO `tailScope`, which this
-    // class's @After never cancelled/joined (unlike factoryScope /
-    // defaultTeardownScope). So a live drain — or its
-    // `invokeOnCompletion -> bridgeScope.launch` (a Main-bound scope) foreign-IO
-    // `Dispatchers.Main` READ on non-cancel completion — survived past
-    // `resetMain` and, racing the next test's `setMain`/`resetMain` WRITE,
-    // threw "Dispatchers.Main is used concurrently with setting it"
-    // (TestMainDispatcher.kt:72), attributed to a random victim test.
-    //
-    // This proves the fix's two properties deterministically on the REAL path
-    // (real VM + real repository + real batched tail):
-    //  (a) LEAK EXISTS — an OPEN tail (its umbrella never completes) leaves a
-    //      LIVE drain coroutine on the tail scope; that is precisely the real-IO
-    //      coroutine the un-joined default scope leaked across the test boundary.
-    //  (b) DRAIN WORKS — cancel-THEN-join (the fix's @After shape, run here while
-    //      the rule's Main is STILL installed, i.e. BEFORE resetMain) drains it,
-    //      so no tail coroutine survives to race the next test's setMain.
-    // The drain is an infinite loop, so it must be cancelled FIRST, then joined.
-    @Test
-    fun issue1168AgentTailScopeLeaksALiveDrainThatCancelJoinDrains() {
-        val probeTailScope = CoroutineScope(SupervisorJob() + Dispatchers.IO)
-        try {
-            val repository = AgentConversationRepository(tailScope = probeTailScope)
-            val vm = newVm(agentRepository = repository)
-            vm.attachClientForTest(FakeTmuxClient())
-            val detection = newClaudeDetection()
-            vm.startAgentConversationForTest("%0", detection)
-            // An OPEN tail: FakeSshSession's tail Job is never completed, so the
-            // umbrella never completes and the batched drain keeps looping on
-            // the injected real-IO tailScope — the un-joined leak.
-            val umbrella = vm.startAgentTailForTest(
-                paneId = "%0",
-                session = FakeSshSession(tailJob = Job()),
-                detection = detection,
-                fromLineExclusive = 0L,
-            )
-            assertNotNull("the follow tail must start (umbrella job non-null)", umbrella)
-
-            // (a) the real-IO drain child is ALIVE on the tail scope.
-            val liveChildren = probeTailScope.coroutineContext.job.children
-                .filter { it.isActive }
-                .toList()
-            assertTrue(
-                "expected a live tail-drain coroutine on the real-IO tailScope " +
-                    "(the un-joined #1168 leak), found none",
-                liveChildren.isNotEmpty(),
-            )
-
-            // (b) cancel-THEN-join drains it (the fix's @After shape), while the
-            // rule's Main is still installed — i.e. before any resetMain.
-            probeTailScope.cancel()
-            runBlocking {
-                withTimeoutOrNull(SLOW_FEED_DRAIN_TIMEOUT_MS) {
-                    probeTailScope.coroutineContext.job.children.forEach { it.join() }
-                }
-            }
-            val survivors = probeTailScope.coroutineContext.job.children
-                .filter { it.isActive }
-                .toList()
-            assertTrue(
-                "cancel-then-join must drain every tail coroutine before " +
-                    "resetMain; survivors=$survivors",
-                survivors.isEmpty(),
-            )
-        } finally {
-            probeTailScope.cancel()
-        }
-    }
-
-    @Test
-    fun staleOldAgentLogTailCompletionDoesNotMarkRestartedPaneStale() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        val oldTailJob = Job()
-        val restartedTailJob = Job()
-        vm.startAgentConversationForTest("%0", detection)
-
-        vm.startAgentTailForTest(
-            paneId = "%0",
-            session = FakeSshSession(tailJob = oldTailJob),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-        vm.startAgentTailForTest(
-            paneId = "%0",
-            session = FakeSshSession(tailJob = restartedTailJob),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-
-        oldTailJob.complete()
-        advanceUntilIdle()
-
-        assertEquals(
-            "old tail completion must not stale a pane that already has a newer tail",
-            AgentConversationSyncStatus.Live,
-            vm.agentConversations.value["%0"]!!.syncStatus,
-        )
-
-        restartedTailJob.complete()
-        advanceUntilIdle()
-
-        assertEquals(AgentConversationSyncStatus.Stale, vm.agentConversations.value["%0"]!!.syncStatus)
-    }
-
-    @Test
-    fun failedAgentLogTailMarksConversationLogUnavailableWhileTmuxStaysConnected() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-        val tailJob = Job()
-        vm.startAgentConversationForTest("%0", detection)
-        vm.startAgentTailForTest(
-            paneId = "%0",
-            session = FakeSshSession(tailJob = tailJob),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-
-        tailJob.completeExceptionally(IOException("tail failed"))
-        advanceUntilIdle()
-
-        assertEquals(
-            AgentConversationSyncStatus.LogUnavailable,
-            vm.agentConversations.value["%0"]!!.syncStatus,
-        )
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-    }
-
-    @Test
-    fun disconnectedSshWhenAgentTailStartsMarksConversationUnavailableWithoutCrashing() = runTest(scheduler) {
-        val diagnostics = installRecordingDiagnosticSink()
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-
-        try {
-            vm.startAgentConversationForPaneForTest(
-                paneId = "%0",
-                session = FakeSshSession(
-                    tailFailure = SshException("SSH session is not connected"),
-                ),
-                detection = detection,
-            )
-            advanceUntilIdle()
-
-            val state = vm.agentConversations.value["%0"]
-            assertEquals(AgentConversationSyncStatus.LogUnavailable, state?.syncStatus)
-            assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-            val event = diagnostics.eventsNamed("tmux_agent_conversation_tail_status").single()
-            assertEquals("recoverable", event.category)
-            assertEquals("%0", event.fields["pane"])
-            assertEquals("tail_start_unavailable", event.fields["reason"])
-            assertEquals(AgentConversationSyncStatus.LogUnavailable.name, event.fields["status"])
-        } finally {
-            diagnostics.close()
-        }
-    }
-
-    @Test
-    fun cancelledAgentConversationStartupReadIsNotSwallowed() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        val detection = newClaudeDetection()
-
-        try {
-            vm.startAgentConversationForPaneForTest(
-                paneId = "%0",
-                session = FakeSshSession(
-                    execFailure = CancellationException("cancelled during line count"),
-                ),
-                detection = detection,
-            )
-            assertTrue("startup cancellation must propagate", false)
-        } catch (e: CancellationException) {
-            assertEquals("cancelled during line count", e.message)
-        }
-        assertTrue(
-            "cancelled startup must not leave a phantom conversation row",
-            vm.agentConversations.value.isEmpty(),
-        )
-    }
-
-    @Test
-    fun retryAgentLogTailForPaneRestartsOneTailAndKeepsTmuxConnected() = runTest(scheduler) {
-        val vm = newVm()
-        val detection = newClaudeDetection()
-        val oldTailJob = Job()
-        val retryTailJob = Job()
-        val recentMtimeSeconds = System.currentTimeMillis() / 1000
-        val retryGate = CompletableDeferred<Unit>()
-        val retrySession = FakeSshSession(
-            tailJob = retryTailJob,
-            execGate = retryGate,
-            detectionOutput = "claude|$recentMtimeSeconds|/work|/home/u/.claude/projects/-work/abc.jsonl\n",
-            processOutput = "123 1 pts/1 node claude\n",
-        )
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%0",
-                    windowId = "@0",
-                    sessionId = "$0",
-                    title = "agent",
-                    paneIndex = 0,
-                    cwd = "/work",
-                    currentCommand = "claude",
-                    paneTty = "/dev/pts/1",
-                ),
-            ),
-        )
-        vm.attachSessionForAgentRetryForTest(retrySession)
-        vm.startAgentConversationForTest("%0", detection)
-        vm.startAgentTailForTest(
-            paneId = "%0",
-            session = FakeSshSession(tailJob = oldTailJob),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-        oldTailJob.complete()
-        advanceUntilIdle()
-
-        assertTrue(vm.retryAgentConversationStreamForPane("%0"))
-        assertEquals(
-            AgentConversationSyncStatus.Retrying,
-            vm.agentConversations.value["%0"]!!.syncStatus,
-        )
-        assertFalse(
-            "duplicate retry must not start a second pane tail",
-            vm.retryAgentConversationStreamForPane("%0"),
-        )
-
-        retryGate.complete(Unit)
-        advanceUntilIdle()
-
-        assertEquals(1, retrySession.tailCalls)
-        assertEquals(AgentConversationSyncStatus.Live, vm.agentConversations.value["%0"]!!.syncStatus)
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-    }
-
-    @Test
-    fun codexAgentLogRetryWithTerminalFloodKeepsConversationAndTerminalConsistent() =
-        runTest(scheduler, timeout = SLOW_FEED_RUN_TEST_TIMEOUT) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        val detection = newCodexDetection()
-        val oldTailJob = Job()
-        val retryTailJob = Job()
-        val retryGate = CompletableDeferred<Unit>()
-        val recentMtimeSeconds = System.currentTimeMillis() / 1000
-        val codexLines = codexTranscriptWithToolFlood(toolResults = 700)
-        val retrySession = FakeSshSession(
-            tailJob = retryTailJob,
-            execGate = retryGate,
-            wcOutput = "${codexLines.size}\n",
-            agentLogLines = codexLines,
-            detectionOutput = "codex|$recentMtimeSeconds|/work|${detection.sourcePath}\n",
-            processOutput = "123 1 pts/1 codex codex --synthetic-overflow\n",
-        )
-        vm.attachClientForTest(client)
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%0",
-                    windowId = "@0",
-                    sessionId = "$0",
-                    title = "codex",
-                    paneIndex = 0,
-                    cwd = "/work",
-                    currentCommand = "codex",
-                    paneTty = "/dev/pts/1",
-                ),
-            ),
-        )
-        vm.attachSessionForAgentRetryForTest(retrySession)
-        vm.startAgentConversationForTest("%0", detection)
-        vm.startAgentTailForTest(
-            paneId = "%0",
-            session = FakeSshSession(tailJob = oldTailJob),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-        oldTailJob.complete()
-        advanceUntilIdle()
-
-        val terminalState = vm.panes.value.single().terminalState
-        val payloads = List(CODEX_SCALE_OUTPUT_CHUNKS, ::codexScaleOutputChunk)
-        val emitted = payloads.sumOf { it.size }
-        assertTrue(
-            "test fixture drift: emitted=$emitted expectedFloor=$CODEX_SCALE_OUTPUT_BYTES",
-            emitted >= CODEX_SCALE_OUTPUT_BYTES,
-        )
-        val observedSideChannelChunks = AtomicInteger(0)
-        val outputCollector = launch {
-            terminalState.output.collect {
-                observedSideChannelChunks.incrementAndGet()
-            }
-        }
-        runCurrent()
-
-        try {
-            assertTrue(vm.retryAgentConversationStreamForPane("%0"))
-            val sender = async {
-                payloads.forEach { bytes ->
-                    client.emittedEvents.emit(ControlEvent.Output("%0", bytes))
-                }
-            }
-            // Issue #1050: the unfixed flood-drain sibling. The old pump here was
-            // bounded only by the virtual `runTest` clock (`withTimeoutOrNull` +
-            // `idle()` + `delay(10)`), so the whole loop elapsed in ~0 wall-clock
-            // time and starved the REAL SshTerminalBridge feed's budgeted
-            // `postDelayed` drain under a contended (CI) JVM -> the flood never
-            // landed and `completed` flaked false. Route through the same
-            // wall-clock-bounded pump #1042 used for codexScale/codexLike.
-            val completed = drainMainLooperUntil { !sender.isActive }
-            if (completed) sender.await() else sender.cancel()
-
-            assertTrue(
-                "Codex-scale terminal output must not stall while Conversation retries",
-                completed,
-            )
-            assertTrue(
-                "Codex-scale terminal output should still publish best-effort side-channel chunks",
-                observedSideChannelChunks.get() > 0,
-            )
-        } finally {
-            outputCollector.cancel()
-        }
-
-        retryGate.complete(Unit)
-        advanceUntilIdle()
-        shadowOf(Looper.getMainLooper()).idle()
-
-        val state = vm.agentConversations.value["%0"]!!
-        val messages = state.events.filterIsInstance<ConversationEvent.Message>()
-        assertEquals(AgentConversationSyncStatus.Live, state.syncStatus)
-        assertTrue(
-            "widened Codex agent-log read must keep the user prompt that is outside the old 200-line tail",
-            messages.any { it.role == ConversationRole.User && it.text == ISSUE_576_CODEX_USER_PROMPT },
-        )
-        assertTrue(
-            "assistant response from the synthetic Codex transcript should also be present",
-            messages.any { it.role == ConversationRole.Assistant && it.text == ISSUE_576_CODEX_ASSISTANT_REPLY },
-        )
-        assertTrue(
-            "Codex conversation overflow must not be classified as a tmux disconnect",
-            vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected,
-        )
-        assertFalse(client.disconnectedSignal.value)
-        assertTrue(
-            "Codex retry must request the widened raw-line window; commands=${retrySession.execCommands}",
-            retrySession.execCommands.any { it.contains("pocketshell agent-log") && it.contains("--tail 1600") },
-        )
-    }
-
-    @Test
-    fun retryAgentLogTailForPaneDetectionFailureReturnsLogUnavailable() = runTest(scheduler) {
-        val vm = newVm()
-        val detection = newClaudeDetection()
-        val retryGate = CompletableDeferred<Unit>()
-        val retrySession = FakeSshSession(execGate = retryGate)
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%0",
-                    windowId = "@0",
-                    sessionId = "$0",
-                    title = "agent",
-                    paneIndex = 0,
-                    cwd = "/work",
-                    currentCommand = "claude",
-                    paneTty = "/dev/pts/1",
-                ),
-            ),
-        )
-        vm.attachSessionForAgentRetryForTest(retrySession)
-        vm.startAgentConversationForTest("%0", detection)
-        val oldTailJob = Job()
-        vm.startAgentTailForTest(
-            paneId = "%0",
-            session = FakeSshSession(tailJob = oldTailJob),
-            detection = detection,
-            fromLineExclusive = 0L,
-        )
-        oldTailJob.complete()
-        advanceUntilIdle()
-
-        assertTrue(vm.retryAgentConversationStreamForPane("%0"))
-        assertEquals(
-            AgentConversationSyncStatus.Retrying,
-            vm.agentConversations.value["%0"]!!.syncStatus,
-        )
-        retryGate.complete(Unit)
-        advanceUntilIdle()
-
-        assertEquals(
-            AgentConversationSyncStatus.LogUnavailable,
-            vm.agentConversations.value["%0"]!!.syncStatus,
-        )
-        assertEquals(0, retrySession.tailCalls)
-        assertTrue(vm.connectionStatus.value is TmuxSessionViewModel.ConnectionStatus.Connected)
-    }
-
     // ─── Issue #186: per-window agent detection state ──────────────────
     //
     // Detection is per-pane (and therefore per-window for the simple
